@@ -1,8 +1,7 @@
 {
   "name": "apollon_standalone",
-  "version": "2.1.3",
+  "version": "2.3.0",
   "private": true,
-  "main": "index.js",
   "repository": "git@github.com:ls1intum/Apollon_standalone.git",
   "license": "MIT",
   "author": {
@@ -11,19 +10,20 @@
     "url": "https://ase.cit.tum.de/krusche"
   },
   "engines": {
-    "node": ">=18.17.0"
+    "node": ">=22.10.0"
   },
-  "workspaces": {
-    "packages": [
-      "packages/*"
-    ]
-  },
+  "workspaces": [
+    "packages/shared",
+    "packages/server",
+    "packages/webapp"
+  ],
   "scripts": {
-    "start": "concurrently \"npm run start:webapp\" \"npm run start:server\" ",
+    "start": "concurrently \"npm run start:webapp\" \"npm run start:server\"",
     "start:webapp": "npm run start --workspace=webapp",
     "start:server": "npm run start --workspace=server",
-    "build": "concurrently \"npm run build:webapp\" \"npm run build:server\" ",
-    "build:local": "concurrently \"npm run build:webapp:local\" \"npm run build:server\" ",
+    "build": "npm run build:shared && concurrently \"npm run build:webapp\" \"npm run build:server\"",
+    "build:local": "npm run build:shared && concurrently \"npm run build:webapp:local\" \"npm run build:server\"",
+    "build:shared": "npm run build --workspace=shared",
     "build:webapp": "npm run build --workspace=webapp",
     "build:webapp:local": "npm run build:local --workspace=webapp",
     "build:server": "npm run build --workspace=server",
@@ -34,11 +34,6 @@
     "prettier:write": "prettier --write 'packages/**/{main,tests}/**/*.{json,ts,tsx,js,jsx,css,scss,html}'",
     "update": "ncu -i --deep --format group"
   },
-  "devDependencies": {
-    "@stylelint/postcss-css-in-js": "0.38.0",
-    "concurrently": "8.2.2",
-    "postcss": "8.4.38"
-  },
   "overrides": {
     "semver": "7.5.4",
     "word-wrap": "1.2.5",
@@ -47,6 +42,23 @@
     "crypto-js": "4.2.0"
   },
   "dependencies": {
-    "fast-json-patch": "3.1.1"
+    "@ls1intum/apollon": "3.3.15"
+  },
+  "devDependencies": {
+    "concurrently": "9.1.0",
+    "typescript": "5.6.3",
+    "eslint": "9.15.0",
+    "prettier": "3.3.3",
+    "@typescript-eslint/eslint-plugin": "8.15.0",
+    "@typescript-eslint/parser": "8.15.0",
+    "stylelint": "16.3.1",
+    "stylelint-config-recommended": "14.0.1",
+    "style-loader": "4.0.0",
+    "css-loader": "7.1.2",
+    "ts-loader": "9.5.1",
+    "webpack": "5.96.1",
+    "webpack-cli": "5.1.4",
+    "webpack-dev-server": "5.1.0",
+    "webpack-merge": "6.0.1"
   }
 }
