@@ -1,49 +1,42 @@
 {
   "name": "server",
-  "version": "2.1.3",
+  "version": "1.0.0",
+  "private": true,
   "license": "MIT",
   "scripts": {
     "clean": "rm -rf ../../build/server",
     "start": "ts-node-dev src/main/server.ts",
     "build": "webpack --config webpack/webpack.config.js",
-    "lint": "eslint -c .eslintrc.js src/main"
+    "lint": "eslint -c .eslintrc.js \"**/*.{ts,tsx}\""
   },
   "engines": {
-    "node": ">=18.17.0"
+    "node": ">=22.10.0"
   },
   "dependencies": {
-    "@ls1intum/apollon": "3.3.14",
-    "@sentry/node": "7.110.1",
+    "@sentry/node": "8.38.0",
     "audit-debounce": "0.1.1",
-    "body-parser": "1.20.2",
-    "canvas": "2.11.2",
+    "body-parser": "1.20.3",
+    "canvas": "^2.11.2",
     "cors": "2.8.5",
-    "express": "4.19.2",
-    "global-jsdom": "9.2.0",
-    "jsdom": "23.2.0",
+    "express": "4.21.1",
+    "fast-json-patch": "3.1.1",
     "ms": "2.1.3",
-    "pdfmake": "0.2.10",
-    "redis": "4.6.13",
+    "pdfmake": "0.2.15",
+    "redis": "4.7.0",
     "rxjs": "7.8.1",
-    "shared": "2.1.3"
+    "shared": "*"
   },
   "devDependencies": {
     "@types/cors": "2.8.17",
     "@types/express": "4.17.21",
     "@types/ms": "0.7.34",
-    "@types/node": "20.12.7",
+    "@types/node": "22.9.0",
     "@types/pdfmake": "0.2.9",
-    "@typescript-eslint/eslint-plugin": "6.21.0",
-    "@typescript-eslint/parser": "6.21.0",
-    "copy-webpack-plugin": "11.0.0",
-    "eslint": "8.57.0",
-    "eslint-config-prettier": "9.1.0",
-    "eslint-plugin-jsdoc": "46.10.1",
-    "eslint-plugin-prefer-arrow": "1.2.3",
-    "eslint-plugin-prettier": "5.1.3",
-    "eslint-plugin-react": "7.34.1",
+    "global-jsdom": "25.0.0",
+    "jsdom": "25.0.1",
+    "node-loader": "^2.1.0",
     "ts-node": "10.9.2",
     "ts-node-dev": "2.0.0",
-    "typescript": "5.4.5"
+    "webpack-node-externals": "^3.0.0"
   }
 }
