@@ -1,14 +1,12 @@
 {
   "name": "shared",
-  "version": "2.1.3",
-  "main": "index.js",
-  "license": "MIT",
-  "dependencies": {
-    "@ls1intum/apollon": "3.3.14",
-    "moment": "2.30.1",
-    "ws": "8.17.1"
+  "version": "1.0.0",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "scripts": {
+    "build": "tsc"
   },
   "engines": {
-    "node": ">=18.17.0"
+    "node": ">=22.10.0"
   }
 }
