@@ -1,85 +1,54 @@
 {
   "name": "webapp",
-  "version": "2.1.3",
+  "version": "1.0.0",
   "private": true,
   "scripts": {
     "clean": "rm -rf ../../build/webapp",
     "start": "webpack serve --config ./webpack/webpack.dev.js",
-    "lint": "npm run lint:ts && npm run lint:css",
-    "lint:ts": "eslint -c .eslintrc.js ./src/main",
-    "lint:css": "stylelint 'src/**/*.ts'",
+    "lint": "eslint -c .eslintrc.js \"**/*.{ts,tsx}\"",
+    "lint:ts": "eslint -c .eslintrc.js \"/src/main\"",
+    "lint:css": "stylelint \"src/**/*.ts\"",
     "build": "npm run clean && webpack --config ./webpack/webpack.prod.js",
     "build:local": "npm run clean && DEPLOYMENT_URL=http://localhost:8080 webpack --config ./webpack/webpack.prod.js"
   },
   "engines": {
-    "node": ">=18.17.0"
+    "node": ">=22.10.0"
   },
   "dependencies": {
-    "@ls1intum/apollon": "3.3.14",
-    "@sentry/react": "7.110.1",
+    "@reduxjs/toolkit": "2.2.7",
+    "@sentry/react": "8.38.0",
     "bootstrap": "5.3.3",
-    "moment": "2.30.1",
-    "postcss-syntax": "0.36.2",
-    "posthog-js": "1.122.0",
+    "posthog-js": "1.186.3",
     "react": "18.2.0",
-    "react-bootstrap": "2.10.2",
+    "react-bootstrap": "2.10.5",
     "react-bootstrap-icons": "1.11.4",
     "react-dom": "18.2.0",
     "react-redux": "8.1.3",
     "react-router-dom": "6.22.3",
-    "redux": "4.2.1",
-    "redux-observable": "2.0.0",
-    "rxjs": "7.8.1",
-    "shared": "2.1.3",
-    "styled-components": "6.1.8",
-    "tslib": "2.6.2",
-    "typesafe-actions": "5.1.0",
+    "shared": "*",
+    "styled-components": "6.1.13",
+    "stylis": "4.3.4",
     "uuid": "9.0.1",
-    "websocket": "1.0.34"
-  },
-  "overrides": {
-    "semver": "7.5.4",
-    "word-wrap": "1.2.5",
-    "tough-cookie": "4.1.3",
-    "postcss": "8.4.38",
-    "crypto-js": "4.2.0"
+    "websocket": "1.0.35"
   },
   "devDependencies": {
     "@stylelint/postcss-css-in-js": "0.38.0",
     "@svgr/webpack": "8.1.0",
     "@types/react": "18.2.79",
-    "@types/react-bootstrap": "0.32.36",
+    "@types/react-bootstrap": "0.32.37",
     "@types/react-dom": "18.2.25",
-    "@types/react-redux": "7.1.33",
+    "@types/react-redux": "^7.1.34",
     "@types/styled-components": "5.1.34",
-    "@types/uuid": "9.0.8",
-    "@typescript-eslint/eslint-plugin": "6.21.0",
-    "@typescript-eslint/parser": "6.21.0",
+    "@types/uuid": "10.0.0",
+    "@types/websocket": "^1.0.10",
     "circular-dependency-plugin": "5.2.2",
     "copy-webpack-plugin": "11.0.0",
     "css-loader": "6.11.0",
     "ejs-loader": "0.5.0",
-    "eslint": "8.57.0",
-    "eslint-config-prettier": "9.1.0",
-    "eslint-plugin-jsdoc": "46.10.1",
-    "eslint-plugin-prefer-arrow": "1.2.3",
-    "eslint-plugin-prettier": "5.1.3",
-    "eslint-plugin-react": "7.34.1",
     "fork-ts-checker-webpack-plugin": "9.0.2",
-    "html-webpack-plugin": "5.6.0",
-    "postcss": "8.4.38",
-    "prettier": "3.2.5",
+    "html-webpack-plugin": "5.6.3",
     "react-toastify": "9.1.3",
-    "style-loader": "3.3.4",
-    "stylelint": "16.3.1",
-    "stylelint-config-recommended": "14.0.0",
     "stylelint-config-styled-components": "0.1.1",
-    "stylelint-processor-styled-components": "1.10.0",
-    "ts-loader": "9.5.1",
-    "typescript": "5.4.5",
-    "webpack": "5.91.0",
-    "webpack-cli": "5.1.4",
-    "webpack-dev-server": "^4.15.2",
-    "webpack-merge": "5.10.0"
+    "stylelint-processor-styled-components": "1.10.0"
   }
 }
