@@ -1,37 +1,36 @@
-/* eslint-env node */
-module.exports = {
-  extends: ['eslint:recommended', 'plugin:@typescript-eslint/recommended'],
-  parser: '@typescript-eslint/parser',
-  plugins: ['@typescript-eslint'],
-  root: true,
-  rules: {
-    'no-constant-condition': 'off',
-    'no-empty': 'off',
-    'no-extra-boolean-cast': 'off',
-    'no-prototype-builtins': 'off',
-    'no-useless-escape': 'off',
-    'prefer-const': 'off',
-    'no-case-declarations': 'off',
-    '@typescript-eslint/ban-ts-comment': 'off',
-    '@typescript-eslint/ban-types': 'off',
-    '@typescript-eslint/no-explicit-any': 'off',
-    '@typescript-eslint/no-namespace': 'off',
-    '@typescript-eslint/no-this-alias': 'off',
-    '@typescript-eslint/no-unsafe-declaration-merging': 'off',
-    '@typescript-eslint/no-unused-vars': 'off',
-    '@typescript-eslint/no-var-requires': 'off',
-    '@typescript-eslint/quotes': [
-      'error',
-      'single',
-      {
-        'avoidEscape': true,
-        'allowTemplateLiterals': true
-      }
-    ]
-  },
-  overrides: [
-    {
-      files: ['*.ts', '*.tsx'],
+const typescriptEslintPlugin = require('@typescript-eslint/eslint-plugin')
+const typescriptParser = require('@typescript-eslint/parser')
+
+module.exports = [
+  {
+    languageOptions: {
+      parser: typescriptParser,
+      parserOptions: {
+        ecmaVersion: 'latest',
+        sourceType: 'module',
+      },
     },
-  ],
-};
+    plugins: {
+      '@typescript-eslint': typescriptEslintPlugin,
+    },
+    rules: {
+      'no-constant-condition': 'warn', 
+      'no-empty': 'warn', 
+      'prefer-const': 'warn',
+
+      '@typescript-eslint/no-explicit-any': 'warn', 
+      '@typescript-eslint/no-unused-vars': [
+        'warn',
+        { argsIgnorePattern: '^_' }
+      ],
+      '@typescript-eslint/ban-ts-comment': [
+        'warn',
+        { 'ts-ignore': 'allow-with-description' } 
+      ],
+
+      '@typescript-eslint/ban-types': 'off', 
+      '@typescript-eslint/no-namespace': 'off', 
+    },
+    files: ['**/*.ts', '**/*.tsx'],
+  },
+]
