@@ -1,6 +1,8 @@
-import pjson from '../../package.json';
+import {
+  apollonLibraryVersion,
+  appVersion,
+  apollonStandaloneRepositoryLink,
+  apollonLibraryRepositoryLink,
+} from '../../../../root_config';
 
-export const appVersion = pjson.version;
-export const apollonStandaloneRepositoryLink = 'https://github.com/ls1intum/Apollon_standalone';
-export const apollonLibraryVersion = pjson.dependencies['@ls1intum/apollon'];
-export const apollonLibraryRepositoryLink = 'https://github.com/ls1intum/Apollon';
+export { apollonLibraryVersion, appVersion, apollonStandaloneRepositoryLink, apollonLibraryRepositoryLink };
