@@ -1,26 +1,18 @@
-import React, { ChangeEvent, Component, ComponentType } from 'react';
+import React, { ChangeEvent, useEffect, useState } from 'react';
 import { Nav, Navbar } from 'react-bootstrap';
 import { FileMenu } from './menues/file-menu';
 import { HelpMenu } from './menues/help-menu';
 import { ThemeSwitcherMenu } from './menues/theme-switcher-menu';
-import { connect, ConnectedComponent } from 'react-redux';
-import { ApplicationState } from '../store/application-state';
 import styled from 'styled-components';
-import { DiagramRepository } from '../../services/diagram/diagram-repository';
 import { appVersion } from '../../application-constants';
-import { Diagram } from '../../services/diagram/diagram-types';
 import { APPLICATION_SERVER_VERSION } from '../../constant';
-import { ModalRepository } from '../../services/modal/modal-repository';
 import { ModalContentType } from '../modals/application-modal-types';
 import { ConnectClientsComponent } from './connected-clients-component';
-import { Collaborator } from 'shared/src/main/collaborator-dto';
-
-type OwnProps = {};
-
-type StateProps = {
-  diagram: Diagram | null;
-  collaborators: Collaborator[];
-};
+import { useAppDispatch, useAppSelector } from '../store/hooks';
+import { updateDiagramThunk } from '../../services/diagram/diagramSlice';
+import { showModal } from '../../services/modal/modalSlice';
+import { LayoutTextSidebarReverse } from 'react-bootstrap-icons';
+import { selectDisplaySidebar, toggleSidebar } from '../../services/version-management/versionManagementSlice';
 
 const DiagramTitle = styled.input`
   font-size: x-large;
@@ -36,96 +28,80 @@ const ApplicationVersion = styled.span`
   margin-right: 10px;
 `;
 
-type DispatchProps = {
-  updateDiagram: typeof DiagramRepository.updateDiagram;
-  openModal: typeof ModalRepository.showModal;
-};
+const MainContent = styled.div<{ $isSidebarOpen: boolean }>`
+  transition: margin-right 0.3s ease;
+  margin-right: ${(props) => (props.$isSidebarOpen ? '250px' : '0')}; /* Adjust based on sidebar width */
+`;
 
-type Props = OwnProps & StateProps & DispatchProps;
+export const ApplicationBar: React.FC = () => {
+  const dispatch = useAppDispatch();
+  const { diagram } = useAppSelector((state) => state.diagram);
+  const [diagramTitle, setDiagramTitle] = useState<string>(diagram?.title || '');
+  const isSidebarOpen = useAppSelector(selectDisplaySidebar);
+  const urlPath = window.location.pathname;
+  const tokenInUrl = urlPath.substring(1); // This removes the leading "/"
 
-const enhance = connect<StateProps, DispatchProps, OwnProps, ApplicationState>(
-  (state) => {
-    return {
-      diagram: state.diagram,
-      collaborators: state.share.collaborators,
-    };
-  },
-  {
-    updateDiagram: DiagramRepository.updateDiagram,
-    openModal: ModalRepository.showModal,
-  },
-);
+  useEffect(() => {
+    if (diagram?.title) {
+      setDiagramTitle(diagram.title);
+    }
+  }, [diagram?.title]);
 
-type State = { diagramTitle: string };
-
-const getInitialState = (props: Props): State => {
-  return {
-    diagramTitle: props.diagram?.title ? props.diagram.title : '',
+  const changeDiagramTitlePreview = (event: ChangeEvent<HTMLInputElement>) => {
+    setDiagramTitle(event.target.value);
   };
+
+  const changeDiagramTitleApplicationState = () => {
+    if (diagram) {
+      dispatch(updateDiagramThunk({ title: diagramTitle }));
+    }
+  };
+
+  const handleOpenModal = () => {
+    dispatch(showModal({ type: ModalContentType.ShareModal, size: 'lg' }));
+  };
+
+  return (
+    <MainContent $isSidebarOpen={isSidebarOpen}>
+      <Navbar className="navbar" variant="dark" expand="lg">
+        <Navbar.Brand>
+          <img alt="" src="images/logo.png" width="60" height="30" className="d-inline-block align-top" />{' '}
+          <span className="fw-bold ms-2">Apollon</span>
+        </Navbar.Brand>
+        <ApplicationVersion>{appVersion}</ApplicationVersion>
+        <Navbar.Toggle aria-controls="basic-navbar-nav" />
+        <Navbar.Collapse id="basic-navbar-nav">
+          <Nav className="me-auto">
+            <FileMenu />
+            {/*<ViewMenu />*/}
+            {APPLICATION_SERVER_VERSION && (
+              <Nav.Item>
+                <Nav.Link onClick={handleOpenModal}>Share</Nav.Link>
+              </Nav.Item>
+            )}
+            <HelpMenu />
+            <DiagramTitle
+              type="text"
+              value={diagramTitle}
+              onChange={changeDiagramTitlePreview}
+              onBlur={changeDiagramTitleApplicationState}
+            />
+          </Nav>
+        </Navbar.Collapse>
+        {!tokenInUrl && (
+          <Nav.Item
+            onClick={() => {
+              dispatch(toggleSidebar());
+            }}
+          >
+            <div className="sidebar-toggle me-lg-4">
+              <LayoutTextSidebarReverse size={20} color="#AEB1B5" />
+            </div>
+          </Nav.Item>
+        )}
+        {tokenInUrl && <ConnectClientsComponent />}
+        <ThemeSwitcherMenu />
+      </Navbar>
+    </MainContent>
+  );
 };
-
-class ApplicationBarComponent extends Component<Props, State> {
-  state = getInitialState(this.props);
-
-  constructor(props: Props) {
-    super(props);
-    this.changeDiagramTitlePreview = this.changeDiagramTitlePreview.bind(this);
-    this.changeDiagramTitleApplicationState = this.changeDiagramTitleApplicationState.bind(this);
-  }
-
-  componentDidUpdate(prevProps: Readonly<Props>, prevState: Readonly<State>, snapshot?: any) {
-    if (this.props.diagram && prevProps.diagram?.title !== this.props.diagram?.title) {
-      this.setState({ diagramTitle: this.props.diagram.title });
-    }
-  }
-
-  changeDiagramTitlePreview(event: ChangeEvent<HTMLInputElement>) {
-    // changes only diagram title of this component not in global state, this happens on blur
-    this.setState({ diagramTitle: event.target.value });
-  }
-
-  changeDiagramTitleApplicationState(event: ChangeEvent<HTMLInputElement>) {
-    if (this.props.diagram) {
-      this.props.updateDiagram({ title: this.state.diagramTitle });
-    }
-  }
-
-  render() {
-    return (
-      <>
-        <Navbar className="navbar" variant="dark" expand="lg">
-          <Navbar.Brand>
-            <img alt="" src="images/logo.png" width="60" height="30" className="d-inline-block align-top" />{' '}
-            <span className="fw-bold ms-2">Apollon</span>
-          </Navbar.Brand>
-          <ApplicationVersion>{appVersion}</ApplicationVersion>
-          <Navbar.Toggle aria-controls="basic-navbar-nav" />
-          <Navbar.Collapse id="basic-navbar-nav">
-            <Nav className="me-auto">
-              <FileMenu />
-              {/*<ViewMenu />*/}
-              {APPLICATION_SERVER_VERSION && (
-                <Nav.Item>
-                  <Nav.Link onClick={(event: any) => this.props.openModal(ModalContentType.ShareModal, 'lg')}>
-                    Share
-                  </Nav.Link>
-                </Nav.Item>
-              )}
-              <HelpMenu />
-              <DiagramTitle
-                type="text"
-                value={this.state.diagramTitle}
-                onChange={this.changeDiagramTitlePreview}
-                onBlur={this.changeDiagramTitleApplicationState}
-              />
-            </Nav>
-            <ConnectClientsComponent collaborators={this.props.collaborators} />
-          </Navbar.Collapse>
-          <ThemeSwitcherMenu />
-        </Navbar>
-      </>
-    );
-  }
-}
-
-export const ApplicationBar: ConnectedComponent<ComponentType<Props>, OwnProps> = enhance(ApplicationBarComponent);
