@@ -1,10 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import { Collaborator } from 'shared/src/main/collaborator-dto';
-
-type Props = {
-  collaborators: Collaborator[];
-};
+import { useAppSelector } from '../store/hooks';
 
 export const NameContainer = styled.div`
   color: #353d47;
@@ -28,17 +24,23 @@ export const Container = styled.div`
   margin-right: 25px;
 `;
 
-export function ConnectClientsComponent(props: Props) {
-  const { collaborators } = props;
-  const elements = collaborators.map((collaborator) => (
-    <NameContainer
-      style={{ backgroundColor: collaborator?.color }}
-      key={collaborator?.name + '_' + collaborator?.color}
-      title={collaborator?.name || 'Anonymous'}
-    >
-      {collaborator ? collaborator.name?.substring(0, 1) : 'A'}
-    </NameContainer>
-  ));
+export function ConnectClientsComponent() {
+  const collaborators = useAppSelector((state) => state.share.collaborators);
+  const { name } = useAppSelector((state) => state.share.userCollaborationData);
+
+  if (!name || collaborators.length === 0) return;
+
+  const elements = collaborators
+    .filter((c) => c != null && c?.name !== name)
+    .map((collaborator) => (
+      <NameContainer
+        style={{ backgroundColor: collaborator?.color }}
+        key={collaborator?.name + '_' + collaborator?.color}
+        title={collaborator?.name || 'Anonymous'}
+      >
+        {collaborator ? collaborator.name?.substring(0, 1) : 'A'}
+      </NameContainer>
+    ));
 
   return <Container>{elements}</Container>;
 }
