@@ -1,5 +1,4 @@
-import { ComponentType } from 'react';
-import { ModalContentType } from './application-modal-types';
+import { ModalContentProps, ModalContentType } from './application-modal-types';
 import { HelpModelingModal } from './help-modeling-modal/help-modeling-modal';
 import { ImportDiagramModal } from './import-diagram-modal/import-diagram-modal';
 import { InformationModal } from './information-modal/information-modal';
@@ -8,8 +7,12 @@ import { CreateDiagramModal } from './create-diagram-modal/create-diagram-modal'
 import { CreateFromTemplateModal } from './create-diagram-from-template-modal/create-from-template-modal';
 import { ShareModal } from './share-modal/share-modal';
 import { CollaborationModal } from './collaboration-modal/collaboration-modal';
+import { DeleteVersionModal } from './delete-version-modal/delete-version-modal';
+import { RestoreVersionModal } from './restore-version-modal/restore-version-modal';
+import { EditVersionModal } from './edit-version-info-modal/edit-version-info-modal';
+import { CreateVersionModal } from './create-version-modal/create-version-modal';
 
-export const ApplicationModalContent: { [key in ModalContentType]: ComponentType<any> } = {
+export const ApplicationModalContent: { [key in ModalContentType]: React.FC<ModalContentProps> } = {
   [ModalContentType.HelpModelingModal]: HelpModelingModal,
   [ModalContentType.ImportDiagramModal]: ImportDiagramModal,
   [ModalContentType.InformationModal]: InformationModal,
@@ -18,4 +21,8 @@ export const ApplicationModalContent: { [key in ModalContentType]: ComponentType
   [ModalContentType.CreateDiagramFromTemplateModal]: CreateFromTemplateModal,
   [ModalContentType.ShareModal]: ShareModal,
   [ModalContentType.CollaborationModal]: CollaborationModal,
+  [ModalContentType.DeleteVersionModal]: DeleteVersionModal,
+  [ModalContentType.RestoreVersionModal]: RestoreVersionModal,
+  [ModalContentType.EditVersionInfoModal]: EditVersionModal,
+  [ModalContentType.CreateVersionModal]: CreateVersionModal,
 };
