@@ -1,85 +1,45 @@
-import React, { Component, ComponentClass } from 'react';
+import React, { useContext } from 'react';
 import { Modal } from 'react-bootstrap';
 import { LocalStorageRepository } from '../../../services/local-storage/local-storage-repository';
-import { compose } from 'redux';
-import { withApollonEditor } from '../../apollon-editor-component/with-apollon-editor';
-import { connect } from 'react-redux';
-import { ApplicationState } from '../../store/application-state';
 import { LocalStorageDiagramListItem } from '../../../services/local-storage/local-storage-types';
 import { LoadDiagramContent } from './load-diagram-content';
+import { useAppDispatch, useAppSelector } from '../../store/hooks';
+import { useLocalStorage } from '../../../services/local-storage/useLocalStorage';
 import { ModalContentProps } from '../application-modal-types';
+import { loadDiagram } from '../../../services/diagram/diagramSlice';
+import { ApollonEditorContext } from '../../apollon-editor-component/apollon-editor-context';
+import { useNavigate } from 'react-router-dom';
+import { setPreviewedDiagramIndex } from '../../../services/version-management/versionManagementSlice';
 
-type OwnProps = {} & ModalContentProps;
-
-type State = {};
-
-type DispatchProps = {
-  load: typeof LocalStorageRepository.load;
-};
-
-type StateProps = {
-  currentDiagramId?: string;
-};
-
-type Props = StateProps & DispatchProps & OwnProps;
-
-const enhance = compose<ComponentClass<OwnProps>>(
-  withApollonEditor,
-  connect<StateProps, DispatchProps, Props, ApplicationState>(
-    (state) => ({
-      currentDiagramId: state.diagram?.id,
-    }),
-    {
-      load: LocalStorageRepository.load,
-    },
-  ),
-);
-
-const getInitialState = (): State => {
-  return {
-    selectedDiagramId: undefined,
-  };
-};
-
-class LoadDiagramModalComponent extends Component<Props, State> {
-  state = getInitialState();
-
-  constructor(props: Props) {
-    super(props);
-    this.loadDiagram = this.loadDiagram.bind(this);
-  }
-
-  getSavedDiagrams(): LocalStorageDiagramListItem[] {
-    // load localStorageList
+export const LoadDiagramModal: React.FC<ModalContentProps> = ({ close }) => {
+  const { diagram } = useAppSelector((state) => state.diagram);
+  const dispatch = useAppDispatch();
+  const fromlocalStorage = useLocalStorage();
+  const editorContext = useContext(ApollonEditorContext);
+  const getSavedDiagrams = (): LocalStorageDiagramListItem[] => {
     const localDiagrams = LocalStorageRepository.getStoredDiagrams();
-    // return all diagrams, but the current displayed diagram
-    return localDiagrams.filter((storedDiagram) => storedDiagram.id !== this.props.currentDiagramId);
-  }
-
-  handleClose = () => {
-    this.props.close();
-    this.setState(getInitialState());
+    return localDiagrams.filter((storedDiagram) => storedDiagram.id !== diagram?.id);
   };
+  const navigate = useNavigate();
 
-  loadDiagram = (id: string) => {
-    if (id) {
-      this.props.load(id);
+  const onSelect = (id: string) => {
+    const loadedDiagram = fromlocalStorage(id);
+    if (loadedDiagram && loadedDiagram.model && editorContext?.editor) {
+      dispatch(loadDiagram(loadedDiagram));
+      dispatch(setPreviewedDiagramIndex(-1));
+      navigate('/', { relative: 'path' });
     }
-    this.handleClose();
+    close();
   };
 
-  render() {
-    return (
-      <>
-        <Modal.Header closeButton>
-          <Modal.Title>Load Diagram</Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-          <LoadDiagramContent diagrams={this.getSavedDiagrams()} onSelect={this.loadDiagram} />
-        </Modal.Body>
-      </>
-    );
-  }
-}
-
-export const LoadDiagramModal = enhance(LoadDiagramModalComponent);
+  return (
+    <>
+      <Modal.Header closeButton>
+        <Modal.Title>Load Diagram</Modal.Title>
+      </Modal.Header>
+      <Modal.Body>
+        <LoadDiagramContent diagrams={getSavedDiagrams()} onSelect={onSelect} />
+      </Modal.Body>
+    </>
+  );
+};
