@@ -1,7 +1,6 @@
-import { CollaboratorType } from 'shared/src/main/collaborator-dto';
-import { SelectionChangeType } from 'shared/src/main/selection-dto';
-import { Diagram } from '../services/diagram/diagram-types';
+import { CollaboratorType, SelectionChangeType } from 'shared';
 import { Patch } from '@ls1intum/apollon';
+import { Diagram } from '../services/diagram/diagramSlice';
 
 export type CollaborationMessage = {
   diagram?: Diagram;
